requirements.txt
python-binance
ta
pandas
numpy
